refactor : domains 패키지 추가 및 배치
