+++ src/main/java/bridge/domains/BridgeMaker.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domains;
 
 import java.util.List;
